# docs/openapi/mnemonic-v1.yaml
openapi: 3.1.0
info:
  title: Mnemonic Admin REST API
  description: |
    Admin REST API for Mnemonic, a knowledge graph and tooling synchronization server.
    This API manages patterns, agents, and skills. It is consumed by curl or any HTTP
    client for data ingestion and administration.

    The MCP server (port 8081) is a separate protocol for Claude Code integration and
    is not documented in this OpenAPI specification.

    ## Architecture

    Mnemonic is a single Go binary that exposes two protocols:

    - **Admin REST API** (this spec) on port 8080: Read + Write via HTTP
    - **MCP Server** on port 8081: Read-only tools for Claude Code via MCP over Streamable HTTP

    ## Authentication

    MVP: no authentication. The server runs on localhost in a trusted environment.

    Post-MVP: authentication and authorization handled externally by Envoy and OPA.

    ## Error Handling

    All errors follow [RFC 7807 Problem Details](https://tools.ietf.org/html/rfc7807)
    with content type `application/problem+json`. Every response includes an
    `X-Request-ID` correlation header.

    ## Pagination

    List endpoints use cursor-based pagination. Cursors are opaque base64-encoded
    strings. The `has_more` field indicates whether additional pages exist.

    ## Request Correlation

    Every response includes an `X-Request-ID` header. If the client provides this
    header on the request, the server echoes it back. Otherwise the server generates
    a UUID v4.
...
servers:
...
components:
  parameters:
...
    skillNameParam:
      name: name
      in: path
      required: true
      description: Skill unique identifier (lowercase letters, numbers, and hyphens; no leading/trailing hyphens, no consecutive hyphens)
      schema:
        type: string
        pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
        maxLength: 64
      example: mnemonic-sync
...
    cursorParam:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous response
      schema:
        type: string
  headers:
    X-Request-ID:
      description: Correlation ID for request tracing. Echoed from request or generated by server.
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    Location:
      description: URI of the newly created resource
      schema:
        type: string
        format: uri-reference
  schemas:
...
    Skill:
      type: object
      description: |
        Full skill definition including content and child resource counts.
        Aligns with the Agent Skills specification (https://agentskills.io/specification).
        Wire format uses flat fields. The server marshals flat fields to/from a JSONB
        `definition` column internally. The `id` and `crc64` fields are server-generated
        and read-only.
      required:
        - id
        - name
        - description
        - content
        - version
        - crc64
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Server-generated UUID primary key
          readOnly: true
          examples:
            - 550e8400-e29b-41d4-a716-446655440020
        name:
          type: string
          description: |
            Unique identifier. Lowercase letters, numbers, and single hyphens.
            No leading/trailing hyphens, no consecutive hyphens.
          pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
          minLength: 1
          maxLength: 64
          examples:
            - mnemonic-sync
        description:
          type: string
          description: Human-readable description of the skill's purpose
          minLength: 1
          maxLength: 1024
          examples:
            - Synchronize agents and skills from Mnemonic
        content:
          type: string
          minLength: 1
          maxLength: 524288
          description: Skill content (markdown body from SKILL.md, up to 512KB)
        tags:
          type: array
          description: Mnemonic-specific tags for filtering
          maxItems: 20
          items:
            type: string
          default: []
          examples:
            - - sync
              - infrastructure
        license:
          type:
            - string
            - "null"
          maxLength: 255
          description: License identifier (e.g., "MIT", "Apache-2.0")
          examples:
            - MIT
        compatibility:
          type:
            - string
            - "null"
          maxLength: 500
          description: Compatibility notes (e.g., supported platforms, required tools)
          examples:
            - "Claude Code 1.0+"
        metadata:
          type:
            - object
            - "null"
          description: Arbitrary key-value metadata (string keys, string values)
          additionalProperties:
            type: string
          examples:
            - author: "team-platform"
              category: "infrastructure"
        allowed_tools:
          type: array
          description: Tool names the skill is allowed to use (experimental, from Agent Skills spec `allowed-tools`)
          maxItems: 50
          items:
            type: string
          default: []
          examples:
            - - Read
              - Write
              - Bash
        version:
          type: string
          minLength: 1
          maxLength: 50
          examples:
            - "1.0.0"
        files:
          type: object
          description: Summary counts of child resources (scripts, references, assets)
          readOnly: true
          properties:
            scripts_count:
              type: integer
              description: Number of script files attached to this skill
            references_count:
              type: integer
              description: Number of reference files attached to this skill
            assets_count:
              type: integer
              description: Number of asset files attached to this skill
        crc64:
          type: string
          description: CRC-64 checksum of the serialized definition (server-computed, decimal string)
          readOnly: true
          examples:
            - "9876543210987654321"
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    SkillCreate:
      type: object
      description: Request body for creating a new skill
      required:
        - name
        - description
        - content
        - version
      properties:
        name:
          type: string
          pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
          minLength: 1
          maxLength: 64
        description:
          type: string
          minLength: 1
          maxLength: 1024
        content:
          type: string
          minLength: 1
          maxLength: 524288
        tags:
          type: array
          maxItems: 20
          items:
            type: string
          default: []
        license:
          type: string
          maxLength: 255
        compatibility:
          type: string
          maxLength: 500
        metadata:
          type: object
          additionalProperties:
            type: string
        allowed_tools:
          type: array
          maxItems: 50
          items:
            type: string
          default: []
        version:
          type: string
          minLength: 1
          maxLength: 50
    SkillUpdate:
      type: object
      description: Request body for updating a skill (full replacement). Name in body must match path parameter or be omitted.
      required:
        - description
        - content
        - version
      properties:
        name:
          type: string
          description: Must match the path parameter (or be omitted)
          pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
          minLength: 1
          maxLength: 64
        description:
          type: string
          minLength: 1
          maxLength: 1024
        content:
          type: string
          minLength: 1
          maxLength: 524288
        tags:
          type: array
          maxItems: 20
          items:
            type: string
        license:
          type: string
          maxLength: 255
        compatibility:
          type: string
          maxLength: 500
        metadata:
          type: object
          additionalProperties:
            type: string
        allowed_tools:
          type: array
          maxItems: 50
          items:
            type: string
        version:
          type: string
          minLength: 1
          maxLength: 50
    SkillList:
      type: object
      description: Paginated list of skills
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ==================== Skill File Schemas ====================
...
  responses:
    BadRequest:
      description: Validation error or malformed request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/validation-error
            title: Validation Error
            status: 400
            detail: The request body contains invalid fields
            instance: /v1/api/agents
            traceId: 550e8400-e29b-41d4-a716-446655440000
            errors:
              - field: name
                code: INVALID_FORMAT
                message: "Name must match pattern ^[a-z]([a-z0-9](-[a-z0-9])*)*$"
              - field: system_prompt
                code: REQUIRED
                message: System prompt is required
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/not-found
            title: Not Found
            status: 404
            detail: "Agent with name 'unknown-agent' not found"
            instance: /v1/api/agents/unknown-agent
            traceId: 550e8400-e29b-41d4-a716-446655440000
    Conflict:
      description: Resource conflict (e.g., duplicate name)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/conflict
            title: Conflict
            status: 409
            detail: "Agent with name 'go-software-engineer' already exists"
            traceId: 550e8400-e29b-41d4-a716-446655440000
...
    InternalError:
      description: Internal server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/internal-error
            title: Internal Error
            status: 500
            detail: An unexpected error occurred
            traceId: 550e8400-e29b-41d4-a716-446655440000
...
paths:
...
  /v1/api/skills:
    get:
      tags:
        - Skills
      summary: List skills
      description: |
        Returns all skills with full content. The sync use case requires full
        content to write to disk. Default page size is 100.
      operationId: listSkills
      parameters:
        - name: limit
          in: query
          description: Maximum items per page
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
        - $ref: "#/components/parameters/cursorParam"
        - name: tags
          in: query
          description: Comma-separated tags to filter by
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of skills with full content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillList"
              example:
                data:
                  - id: 550e8400-e29b-41d4-a716-446655440020
                    name: mnemonic-sync
                    description: Synchronize agents and skills from Mnemonic
                    content: "---\nname: mnemonic-sync\ndescription: Synchronize...\n---\n\nYou are synchronizing..."
                    tags:
                      - sync
                      - infrastructure
                    license: MIT
                    compatibility: null
                    metadata: null
                    allowed_tools: []
                    version: "1.0.0"
                    files:
                      scripts_count: 1
                      references_count: 0
                      assets_count: 0
                    crc64: "9876543210987654321"
                    created_at: "2026-02-01T10:00:00Z"
                    updated_at: "2026-02-10T14:00:00Z"
                pagination:
                  limit: 100
                  cursor: null
                  next_cursor: null
                  has_more: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Skills
      summary: Create skill
      operationId: createSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillCreate"
            example:
              name: code-review
              description: Multi-agent code review skill
              content: "---\nname: code-review\n---\n\nYou are reviewing code..."
              tags:
                - review
                - quality
              license: MIT
              version: "1.0.0"
      responses:
        "201":
          description: Skill created
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/api/skills/{name}:
    parameters:
      - $ref: "#/components/parameters/skillNameParam"

    get:
      tags:
        - Skills
      summary: Get skill by name
      operationId: getSkill
      responses:
        "200":
          description: Skill details
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
              example:
                id: 550e8400-e29b-41d4-a716-446655440020
                name: mnemonic-sync
                description: Synchronize agents and skills from Mnemonic
                content: "---\nname: mnemonic-sync\n..."
                tags:
                  - sync
                  - infrastructure
                license: MIT
                compatibility: null
                metadata: null
                allowed_tools: []
                version: "1.0.0"
                files:
                  scripts_count: 1
                  references_count: 0
                  assets_count: 0
                crc64: "9876543210987654321"
                created_at: "2026-02-01T10:00:00Z"
                updated_at: "2026-02-10T14:00:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags:
        - Skills
      summary: Update skill
      description: Full replacement. Name in body must match path parameter.
      operationId: updateSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "204":
          description: No Content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags:
        - Skills
      summary: Delete skill
      operationId: deleteSkill
      responses:
        "204":
          description: Skill deleted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ==================== Skill File Endpoints ====================

  # --- Scripts ---
...